Organize chatbot into sub componenets
